backend_service/pyproject.toml
[project]
name = "backend-llm-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.1.1",
    "fastapi==0.104.1",
    "httpx==0.25.2",
    "langchain==0.1.0",
    "langchain-openai==0.0.5",
    "psycopg2-binary==2.9.9",
    "pydantic[email]==2.5.0",
    "pydantic-settings>=2.2.1",
    "pyjwt>=2.10.1",
    "python-dotenv==1.0.0",
    "redis==5.0.1",
    "sqlalchemy==2.0.23",
    "uvicorn[standard]==0.24.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.18.2",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
]
